+++ addOns/ascanrulesAlpha/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Changed
+- The Web Cache Deception scan rule now includes example alert functionality for documentation generation purposes (Issue 6119).
 
 ## [52] - 2025-10-07
 ### Added